Moved Some classes into components (Header, SiteSection) and updated components
