+++ src/components/loaders/search-loader.tsx
@@ -1,10 +1,11 @@
+import SiteSection from "../site-section";
 import BlogItemLoader from "./ui/blog-item";
 import PaginationLoader from "./ui/pagination";
 import SearchLoader from "./ui/search";
 
 export default function SearchResultsLoader({ count = 10 }: { count?: number }){
      return (
-          <section className="scroll-mt-10 min-h-[56vh]" aria-hidden="true">
+          <SiteSection className="scroll-mt-10 min-h-[56vh]" aria-hidden="true">
                <div className="inner-width flex flex-col items-center justify-center gap-6">
                     <SearchLoader/>
                     <div className="space-y-5 w-full">
@@ -12,6 +13,6 @@ export default function SearchResultsLoader({ count = 10 }: { count?: number }){
                          <PaginationLoader/>
                     </div>
                </div>
-          </section>
+          </SiteSection>
      )
 }
