+++ src/components/sections/search.tsx
@@ -4,7 +4,7 @@ import { Search } from "lucide-react"
 import { Input } from "@/components/ui/input"
 import BlogItem from "../blog/item"
 import SiteSection from "../layout/section"
-import { PaginationWithLinks } from "@/components/ui/pagination-with-links"
+import { PaginationWithLinks } from "@/components/pagination-with-links"
 import { Form, FormControl, FormField, FormItem } from "@/components/ui/form"
 import * as z from "zod"
 import { UseFormReturn } from "react-hook-form"
